update make_landmap.py
Fixed issue: prevent to add more smaller area later. It has not been prevented, but it has been made possible without bad consequences on the map as it was before.
Fixed a bug when loading from an existing map. The bug was related to a not proper cleaning. Now it is done properly.
Naming system of layers of BaseDetail class has been simplified.