update make_landmap.py
Edit: do not delete the full step when clicking the previous button, just hide everything. And show it again when clicking the next button. Add a cancel button to cancel the step also in global builders.

To do so, cleandrawables method has been substituted with dhsdrawables method, and the new method TSLbase.editvislayers() is added.
Still something to be fixed: there are some hidden layers that are suddenly shown when going back and forward (should stay always hidden) and the cancel button to GlobalBuilder has to be added.