update make_landmap.py
fixed issues with actions from beforerun or beforegen when going back to a previous step. Now afterclosing is always called to clean stuffs.
Now is possible to add a symbol when going back from a previous step (the underlying yellow shape is added correctly).

Still to do: if the user is gone back to a previous step and want to regenerate a step / add other local mutilgen steps, let the new layers be created in the right position. In some case is mandatory to do this, in others may be useful to let the new layers generated at the top of the layer stack.