update two files, add one
Font existence check added. Now font names are stored in a separate text file make_landmap_fonts, which is read each time the plug-in is launched. Provides message errors if font are not valid and instruction on how to edit it.
Also install.sh has been modified so that the make_landmap_fonts file is copied in the correct position